backport of commit 843c50e8ce815ba233816e06bd68503a771abd93
