+++ internal/lang/types/types.go
@@ -1,0 +1,2 @@
+// Package types contains non-standard cty types used only within Terraform.
+package types
